requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching & Task Queue
redis==5.0.1
celery==5.3.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
slowapi==0.1.9

# Logging
structlog>=24.1.0

# LLM Integration
anthropic>=0.76.0
tenacity>=9.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2
